+++ src/components/container.ts
@@ -1,6 +1,8 @@
-const baseStyle = {
-  maxW: '7xl',
-  px: { base: '4', md: '8' },
-}
+import { defineStyleConfig } from '@chakra-ui/styled-system'
 
-export default { baseStyle }
+export default defineStyleConfig({
+  baseStyle: {
+    maxW: '7xl',
+    px: { base: '4', md: '8' },
+  },
+})